+++ rend3/src/lib.rs
@@ -104,6 +104,7 @@ pub mod managers {
     mod material;
     mod mesh;
     mod object;
+    mod point;
     mod skeleton;
     mod texture;
 
@@ -114,6 +115,7 @@ pub mod managers {
     pub use material::*;
     pub use mesh::*;
     pub use object::*;
+    pub use point::*;
     pub use skeleton::*;
     pub use texture::*;
 }